+++ packages/vite-plugin-angular/src/lib/utils/hmr-candidates.ts
@@ -139,6 +139,7 @@ export function analyzeFileUpdates(
 
       // Check for component class nodes
       const meta = compiler?.getMeta(updatedNode);
+      console.log('meta', stale.fileName, meta);
       if (
         meta?.decorator &&
         (meta as { isComponent?: boolean }).isComponent === true
@@ -200,14 +201,13 @@ export function analyzeFileUpdates(
         }
 
         // Compare component meta decorator object literals
-        if (
-          hasUnsupportedMetaUpdates(
-            staleDecoratorExpression,
-            stale,
-            updatedDecoratorExpression,
-            updated,
-          )
-        ) {
+        const analysis = analyzeMetaUpdates(
+          staleDecoratorExpression,
+          stale,
+          updatedDecoratorExpression,
+          updated,
+        );
+        if (analysis === MetaUpdateAnalysis.Unsupported) {
           return null;
         }
 
@@ -225,7 +225,9 @@ export function analyzeFileUpdates(
         }
 
         // If all previous class checks passed, this class is supported for HMR updates
-        candidates.push(updatedNode);
+        if (analysis === MetaUpdateAnalysis.Supported) {
+          candidates.push(updatedNode);
+        }
         continue;
       }
     }
@@ -244,13 +246,19 @@ export function analyzeFileUpdates(
 /**
  * The set of Angular component metadata fields that are supported by HMR updates.
  */
-const SUPPORTED_FIELDS = new Set([
+const SUPPORTED_FIELD_NAMES = new Set([
   'template',
   'templateUrl',
   'styles',
   'styleUrl',
   'stylesUrl',
 ]);
+
+enum MetaUpdateAnalysis {
+  Supported,
+  Unsupported,
+  None,
+}
 
 /**
  * Analyzes the metadata fields of a decorator call expression for unsupported HMR updates.
@@ -259,24 +267,26 @@ const SUPPORTED_FIELDS = new Set([
  * @param staleSource The source file instance containing the stale call instance.
  * @param updatedCall A call expression instance.
  * @param updatedSource The source file instance containing the updated call instance.
- * @returns true, if unsupported metadata updates are present; false, otherwise.
+ * @returns A MetaUpdateAnalysis enum value.
  */
-function hasUnsupportedMetaUpdates(
+function analyzeMetaUpdates(
   staleCall: ts.CallExpression,
   staleSource: ts.SourceFile,
   updatedCall: ts.CallExpression,
   updatedSource: ts.SourceFile,
-): boolean {
+): MetaUpdateAnalysis {
   const staleObject = staleCall.arguments[0];
   const updatedObject = updatedCall.arguments[0];
+  let hasSupportedUpdate = false;
 
   if (
     !ts.isObjectLiteralExpression(staleObject) ||
     !ts.isObjectLiteralExpression(updatedObject)
   ) {
-    return true;
+    return MetaUpdateAnalysis.Unsupported;
   }
 
+  const supportedFields = new Map<string, ts.Node>();
   const unsupportedFields: ts.Node[] = [];
 
   for (const property of staleObject.properties) {
@@ -285,11 +295,12 @@ function hasUnsupportedMetaUpdates(
       ts.isComputedPropertyName(property.name)
     ) {
       // Unsupported object literal property
-      return true;
+      return MetaUpdateAnalysis.Unsupported;
     }
 
     const name = property.name.text;
-    if (SUPPORTED_FIELDS.has(name)) {
+    if (SUPPORTED_FIELD_NAMES.has(name)) {
+      supportedFields.set(name, property.initializer);
       continue;
     }
 
@@ -303,11 +314,26 @@ function hasUnsupportedMetaUpdates(
       ts.isComputedPropertyName(property.name)
     ) {
       // Unsupported object literal property
-      return true;
+      return MetaUpdateAnalysis.Unsupported;
     }
 
     const name = property.name.text;
-    if (SUPPORTED_FIELDS.has(name)) {
+    if (SUPPORTED_FIELD_NAMES.has(name)) {
+      const staleInitializer = supportedFields.get(name);
+      // If the supported field was added or has its content changed, there has been a supported update
+      if (
+        !staleInitializer ||
+        !equalRangeText(
+          property.initializer,
+          updatedSource,
+          staleInitializer,
+          staleSource,
+        )
+      ) {
+        hasSupportedUpdate = true;
+      }
+      // Remove the field entry to allow tracking removed fields
+      supportedFields.delete(name);
       continue;
     }
 
@@ -320,11 +346,20 @@ function hasUnsupportedMetaUpdates(
         staleSource,
       )
     ) {
-      return true;
+      return MetaUpdateAnalysis.Unsupported;
     }
   }
 
-  return i !== unsupportedFields.length;
+  if (i !== unsupportedFields.length) {
+    return MetaUpdateAnalysis.Unsupported;
+  }
+
+  // Any remaining supported field indicates a field removal. This is also considered a supported update.
+  hasSupportedUpdate ||= supportedFields.size > 0;
+
+  return hasSupportedUpdate
+    ? MetaUpdateAnalysis.Supported
+    : MetaUpdateAnalysis.None;
 }
 
 /**