+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [1.16.2-beta.3](https://github.com/analogjs/analog/compare/v1.16.2-beta.2...v1.16.2-beta.3) (2025-05-22)
+
+### Bug Fixes
+
+- **vite-plugin-angular:** fix live reload and move to separate plugin ([#1739](https://github.com/analogjs/analog/issues/1739)) ([0204e30](https://github.com/analogjs/analog/commit/0204e304b2d71d72c1c3cafe5222a3734dc920d2))
+
+## [1.16.2-beta.2](https://github.com/analogjs/analog/compare/v1.16.2-beta.1...v1.16.2-beta.2) (2025-05-17)
+
+### Bug Fixes
+
+- **create-analog:** use latest version of marked w/shiki, fix tailwind support ([#1736](https://github.com/analogjs/analog/issues/1736)) ([3bebec9](https://github.com/analogjs/analog/commit/3bebec99b08fb653eebc2a66c28fe9727d6f2283))
+
+## [1.16.2-beta.1](https://github.com/analogjs/analog/compare/v1.16.1...v1.16.2-beta.1) (2025-05-14)
+
+### Bug Fixes
+
+- **vite-plugin-angular:** normalize version string for pending tasks replacement ([#1731](https://github.com/analogjs/analog/issues/1731)) ([624696c](https://github.com/analogjs/analog/commit/624696c6ce1061951373dc0a94b29710d8997eaa))
+
 ## [1.16.1](https://github.com/analogjs/analog/compare/v1.16.0...v1.16.1) (2025-05-12)
 
 ### Bug Fixes